add forms, use long instead of int for payment type id, add hooks for endpoints, add event page

+++ src/main/java/ee/laus/eventapp/participant/dto/PrivateEntityParticipantDto.java
@@ -5,6 +5,6 @@ public record PrivateEntityParticipantDto(
         String firstName,
         String lastName,
         String details,
-        Integer paymentTypeId
+        Long paymentTypeId
 ) implements ParticipantDto {
 }
